Remove 2 week restriction on all time leaderboard.
I'm not 100% sure why this restriction was added (see the commit linked
below), but I'm removing it for now, as restarting the server has
removed users that expected to have the data there.

Perhaps in the near future the backend can be simplified if the data is
stored on folks' devices instead.